Fix a lot of things, need Testing

+++ src/main/java/com/example/book/service/impl/CommentServiceImpl.java
@@ -1,17 +1,22 @@
 package com.example.book.service.impl;
 
 import com.example.book.config.MessageConfig;
+import com.example.book.constant.RoleType;
+import com.example.book.dto.RequestDTO.CommentRequestDTO;
 import com.example.book.dto.ResponseDTO.Comment.CommentResponseDTO;
 import com.example.book.dto.ResponseDTO.Comment.CommentShortResponseDTO;
+import com.example.book.dto.ResponseDTO.PageResponseDTO;
 import com.example.book.entity.Comment;
+import com.example.book.entity.User;
 import com.example.book.exception.ResourceNotFoundException;
 import com.example.book.repository.CommentRepository;
 import com.example.book.repository.PostRepository;
 import com.example.book.repository.UserRepository;
 import com.example.book.service.CommentService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Service
 public class CommentServiceImpl implements CommentService {
@@ -20,45 +25,55 @@ public class CommentServiceImpl implements CommentService {
     private final UserRepository userRepository;
     private final PostRepository postRepository;
     private final MessageConfig messageConfig;
+    private final UserServiceImpl userServiceImpl;
     private final String POST_NOT_FOUND = "error.post.notfound";
     private final String COMMENT_NOT_FOUND = "error.comment.notfound";
 
-    public CommentServiceImpl(CommentRepository commentRepository, UserRepository userRepository, PostRepository postRepository, MessageConfig messageConfig) {
+    public CommentServiceImpl(CommentRepository commentRepository, UserRepository userRepository, PostRepository postRepository, MessageConfig messageConfig, UserServiceImpl userServiceImpl) {
         this.commentRepository = commentRepository;
         this.userRepository = userRepository;
         this.postRepository = postRepository;
         this.messageConfig = messageConfig;
+        this.userServiceImpl = userServiceImpl;
     }
 
     @Override
-    public CommentShortResponseDTO addComment(Comment comment) {
-        if(!postRepository.existsById(comment.getPost().getPostId())) {
-            throw new ResourceNotFoundException(messageConfig.getMessage(POST_NOT_FOUND, comment.getPost().getPostId()));
+    public CommentShortResponseDTO addComment(Long postId, CommentRequestDTO request) {
+        if(!postRepository.existsById(postId)){
+            throw new ResourceNotFoundException(messageConfig.getMessage(POST_NOT_FOUND, postId));
         }
+        Comment comment = new Comment();
+        comment.setCommentDetail(request.getContent());
         commentRepository.save(comment);
         return convertCommentToShortDTO(comment);
     }
 
     @Override
-    public CommentShortResponseDTO updateComment(Long id, Comment comment) {
-        Optional<Comment> optionalComment = commentRepository.findById(id);
+    public CommentShortResponseDTO updateComment(Long postId, Long commentId, CommentRequestDTO request) {
+        if(!postRepository.existsById(postId)){
+            throw new ResourceNotFoundException(messageConfig.getMessage(POST_NOT_FOUND, postId));
+        }
+        Optional<Comment> optionalComment = commentRepository.findById(commentId);
         if (optionalComment.isPresent()) {
             Comment updatedComment = optionalComment.get();
-            updatedComment.setCommentDetail(comment.getCommentDetail());
-            updatedComment.setLastModifiedDate(comment.getLastModifiedDate());
+            updatedComment.setCommentDetail(request.getContent());
             commentRepository.save(updatedComment);
             return convertCommentToShortDTO(updatedComment);
         } else {
-            throw new ResourceNotFoundException(messageConfig.getMessage(POST_NOT_FOUND, comment.getPost().getPostId()));
+            throw new ResourceNotFoundException(messageConfig.getMessage(COMMENT_NOT_FOUND, commentId));
         }
     }
 
     @Override
-    public List<CommentShortResponseDTO> getComments() {
-        List<Comment> comments = commentRepository.findAll();
-        List<CommentShortResponseDTO> commentShortResponseDTOS = comments.stream()
-                .map(this::convertCommentToShortDTO).toList();
-        return commentShortResponseDTOS;
+    public PageResponseDTO<CommentShortResponseDTO> getComments(Pageable pageable) {
+        Page<Comment> comments = commentRepository.findAll(pageable);
+        Page<CommentShortResponseDTO> commentPage = comments.map(c -> convertCommentToShortDTO(c));
+        return new PageResponseDTO<>(
+                commentPage.getNumber() + 1,
+                commentPage.getNumberOfElements(),
+                commentPage.getTotalPages(),
+                commentPage.getContent()
+        );
     }
 
     @Override
@@ -94,30 +109,31 @@ public class CommentServiceImpl implements CommentService {
                 commentRoots.add(commentNode);
             }
         }
-        //Comparator<CommentResponseDTO> comparator = Comparator.comparing(CommentResponseDTO::getCreatedAt);
-        //commentRoots.sort(comparator.reversed());
+        Comparator<CommentResponseDTO> comparator = Comparator.comparing(CommentResponseDTO::getCreatedAt);
+        commentRoots.sort(comparator.reversed());
         return commentRoots;
     }
 
-    //de the nay thi no se xoa het ca nhung comment con
-    //co the can nhac de neu xoa comment nao thi ghi de ten nguoi dung la deleted, con noi dung comment la comment removed by user
     @Override
     public void deleteComment(Long id){
         Comment commentDeleted = commentRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(messageConfig.getMessage(COMMENT_NOT_FOUND, id)));
-        List<Comment> comments = commentRepository.findAllByParent_CommentId(id);
-        comments.forEach(c -> { if(commentDeleted.getParent() != null)
+        User currentUser = userServiceImpl.getCurrentUser();
+        if(currentUser.getRole().getRoleName().equals(RoleType.ADMIN) ||
+                currentUser.equals(commentDeleted.getUser()) || currentUser.equals(commentDeleted.getPost().getUser())){
+            List<Comment> comments = commentRepository.findAllByParent_CommentId(id);
+            comments.forEach(c -> { if(commentDeleted.getParent() != null)
             {
                 c.setParent(commentDeleted.getParent()); commentRepository.save(c);
             }
             else {c.setParent(null); commentRepository.save(c);}
-        });
-        commentRepository.delete(commentDeleted);
+            });
+            commentRepository.delete(commentDeleted);
+        }
     }
 
     public CommentResponseDTO convertCommentToDTO(Comment comment){
         CommentResponseDTO commentResponse = new CommentResponseDTO();
         commentResponse.setCommentId(comment.getCommentId());
-        commentResponse.setCommentDetail(comment.getCommentDetail());
         commentResponse.setCreatedAt(comment.getCreatedDate());
         commentResponse.setCommentDetail(comment.getCommentDetail());
         commentResponse.setUpdatedAt(comment.getLastModifiedDate());
